Rename dependency to provider
