+++ pkg/auth/provider/provider.go
@@ -1,15 +1,15 @@
-package dependency
+package provider
 
 import (
 	"github.com/skygeario/skygear-server/pkg/core/config"
 	"github.com/skygeario/skygear-server/pkg/core/db"
 )
 
-type AuthDependency struct {
+type AuthProviders struct {
 	DB *db.DBProvider
 }
 
-func (d AuthDependency) Provide(dependencyName string, tConfig config.TenantConfiguration) interface{} {
+func (d AuthProviders) Provide(dependencyName string, tConfig config.TenantConfiguration) interface{} {
 	switch dependencyName {
 	case "DB":
 		return d.DB.Provide(tConfig)
